updated on 2024-03-31 13:26

+++ problems/starry_algo/problem/problem8.cpp
@@ -15,7 +15,7 @@ using ll = long long;
 
 const int N = 1e5;
 int n, p, q;
-int map[N], diff[N], prefix[N];
+ll map[N], diff[N], prefix[N];
 
 void solve() {
     cin >> n >> p >> q;
